+++ core/scripts/version-check.py
@@ -47,6 +47,7 @@ RAVEN_RELEASES = [
     "5.5.1",
     "5.5.2",
     "5.5.3",
+    "5.5.4",
 ]
 
 RAVEN_LATEST = RAVEN_RELEASES[-1]